move execute group to execute underlying module
