+++ sharding-core/sharding-core-execute/src/test/java/org/apache/shardingsphere/sharding/execute/sql/execute/SQLExecuteCallbackTest.java
@@ -20,7 +20,7 @@ package org.apache.shardingsphere.sharding.execute.sql.execute;
 import lombok.SneakyThrows;
 import org.apache.shardingsphere.underlying.executor.constant.ConnectionMode;
 import org.apache.shardingsphere.underlying.common.database.type.DatabaseTypes;
-import org.apache.shardingsphere.sharding.execute.sql.StatementExecuteUnit;
+import org.apache.shardingsphere.underlying.executor.jdbc.StatementExecuteUnit;
 import org.apache.shardingsphere.underlying.executor.context.ExecutionUnit;
 import org.apache.shardingsphere.underlying.executor.context.SQLUnit;
 import org.apache.shardingsphere.underlying.common.database.metadata.DataSourceMetaData;
