+++ shardingsphere-underlying/shardingsphere-executor/src/main/java/org/apache/shardingsphere/underlying/executor/jdbc/connection/StatementOption.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.underlying.executor.connection;
+package org.apache.shardingsphere.underlying.executor.jdbc.connection;
 
 import lombok.AccessLevel;
 import lombok.Getter;
